refactor: consolidate site note screens and  components to same folder
